move to directory api_exposer
