+++ api_exposer/validator.py
@@ -5,7 +5,7 @@ from fastapi import HTTPException
 from matter_server.client.models.node import MatterNode, MatterEndpoint
 from chip.clusters.ClusterObjects import Cluster, ClusterCommand, ClusterObjectFieldDescriptor
 from chip.clusters import Objects
-from nodes import Nodes as ClientNodes
+from api_exposer.nodes import Nodes as ClientNodes
 
 
 def not_found(msg: str) -> None:
