+++ nodes_usage.py
@@ -4,9 +4,9 @@ import asyncio
 import logging
 import argparse
 
-from nodes import Nodes
+from api_exposer.nodes import Nodes
 
-from const import DEFAULT_SERVER_URL
+from api_exposer.const import DEFAULT_SERVER_URL
 
 # Get parsed passed in arguments.
 parser = argparse.ArgumentParser(
